+++ src/App.js
@@ -2,7 +2,12 @@
 
 import axios from "axios";
 import React, { useState, useEffect } from "react";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import {
+  BrowserRouter as Router,
+  Route,
+  Routes,
+  useNavigate,
+} from "react-router-dom";
 import Dashboard from "./pages/dashBoard/dashBoardPage";
 import Search from "./pages/search/searchPage";
 import MainPage from "./pages/main/mainPage";
@@ -19,16 +24,14 @@ function App() {
   });
 
   return (
-    <>
-      <MainPage />
-      <Router>
-        <Routes>
-          <Route path="/dash-board" element={<Dashboard />} />{" "}
-          {/* 대시보드 페이지 */}
-          <Route path="/search" element={<Search />} /> {/* 검색 페이지 */}
-        </Routes>
-      </Router>
-    </>
+    <Router>
+      <Routes>
+        <Route path="/" element={<MainPage />} /> {/* 메인 페이지 */}
+        <Route path="/dash-board" element={<Dashboard />} />{" "}
+        {/* 대시보드 페이지 */}
+        <Route path="/search" element={<Search />} /> {/* 검색 페이지 */}
+      </Routes>
+    </Router>
   );
 }
