+++ src/pages/dashBoard/dashBoardPage.module.css
@@ -1,7 +1,8 @@
 .dashBoard {
   position: relative;
-  width: 100%;
-  height: 100vh;
+  height: 1671px;
+  background-color: #fff;
+  margin-bottom: 100px;
 }
 
 .title {
@@ -15,7 +16,7 @@
 
 .search {
   position: absolute;
-  left: 505px;
+  left: 705px;
   top: 225px;
   width: 424px;
   height: 40px;
@@ -45,7 +46,7 @@
 
 .spaceCreateBtn {
   position: absolute;
-  left: 1232px;
+  left: 1600px;
   top: 225px;
   width: 154px;
   height: 43px;
