+++ src/pages/search/searchPage.module.css
@@ -1,7 +1,8 @@
 .search {
   position: relative;
-  width: 100%;
-  height: 100vh;
+  height: 1671px;
+  background-color: #fff;
+  margin-bottom: 100px;
 }
 
 .title {
@@ -16,7 +17,7 @@
 
 .createSpaceBtn {
   position: absolute;
-  left: 1232px;
+  left: 1600px;
   top: 225px;
   width: 154px;
   height: 43px;
